+++ Sources/QAMenuUIKit/Public/UI/QAMenuUIKitPresenter.swift
@@ -83,8 +83,12 @@ open class QAMenuUIKitPresenter: QAMenuPresenter {
             Logger.verbose("Trying to setup the qa menu window, but the qa menu object is nil")
             return
         }
-        let paneRepresentable = self.ui.retrieve(for: type(of: qaMenu.pane))
-        let uiRepresentation = paneRepresentable.make(for: qaMenu.pane, in: qaMenu)
+        guard let rootPane = qaMenu.pane else {
+            Logger.error("You need to set a root pane before using qa menu!")
+            return
+        }
+        let paneRepresentable = self.ui.retrieve(for: type(of: rootPane))
+        let uiRepresentation = paneRepresentable.make(for: rootPane, in: qaMenu)
 
         let navigationController = UINavigationController(rootViewController: uiRepresentation)